Incorporates some changes for "mainline" EMKeychain into Steven Degutis' pleasantly refactored version. Chief among them, EMKeychain now fills out its ivars using data returned by the Keychain, not from what the caller handed to the class method.
  EMGenericKeychainItem *kcItem = [EMGenericKeychainItem genericKeychainItemForService: @"MyApp" withUsername: nil];

If found, the resulting EMKeychainItem will come into existence with
its username set per the Keychain. Ditto service names and all the
rest on Internet keychain items.

I also fixed some deprecation warnings and added a set of unit
tests. As of this writing, all tests pass on 10.5 (intel/ppc) and on
10.6 (intel), with Xcode 3.1.3 and 3.2 respectively. I haven't tried
it on 10.4, but there's no obvious reason it wouldn't work.